requirements.txt
pandas
python-docx
numpy
wasabi
psycopmlutils
spacy
textdescriptives
yaml